# docs/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Swagger Commune",
        "description": "TODO",
        "contact": {
            "email": "iam@it-yaroslav.ru"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8090"
        }
    ],
...
    "paths": {
...
        "/auth/reset": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Reset passcode and send to enail new one",
                "description": "Reset passcode and send to enail new one",
                "operationId": "resetPasscode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSignUp"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserSignUp": {
                "required": [
                    "email",
                    "nickname"
                ],
                "type": "object",
                "properties": {
                    "nickname": {
                        "$ref": "#/components/schemas/Nickname"
                    },
                    "email": {
                        "$ref": "#/components/schemas/Email"
                    }
                }
            },
...
            "MessageResponse": {
                "required": ["message"],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Something went wrong"
                    }
                }
            },
            "Nickname": {
                "type": "string",
                "example": "SomeNickname123"
            },
            "Email": {
                "description": "Unique email. A passcode will be sent to it",
                "type": "string",
                "example": "example@mail.ru"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}